Home

InsarLOS.jl

InsarLOS.calculate_los_xyz

InsarLOS.create_los_map

InsarLOS.get_los_enu

InsarLOS.orbithermite

InsarLOS.project_enu_to_los

InsarLOS.rot

InsarLOS.rotate_xyz_to_enu

InsarLOS.calculate_los_xyz — Method.

Calculate the line of sight vector (from satellite to ground) in X,Y,Z coordinates at one lat/lon point

source

InsarLOS.create_los_map — Method.

Create a map with 3 layers (E, N, U) of the ENU vectors for every pixel within a demrsc

source

InsarLOS.get_los_enu — Method.

get_los_enu(lat_lons::AbstractArray{<:AbstractFloat, 2}; xyz_los_vecs=nothing, dbfile=nothing, param_dict=nothing)

Calculate the LOS vector in ENU from a lat/lon Array Can also take in the pre-computed xyz los vectors

lat_lons is 2 x N for N different points

source

InsarLOS.orbithermite — Method.

orbithermite - hermite polynomial interpolation of orbits

source

InsarLOS.project_enu_to_los — Function.

Find magnitude of an ENU vector in the LOS direction

Rotates the line of sight vector to ENU coordinates at (lat, lon), then dots with the enu data vector

Args: enu (list[float], ndarray[float]): E,N,U coordinates, either as list of 3, or a (3, k) array of k ENU vectors losvec (ndarray[float]) length 3 line of sight, in XYZ frame lat (float): degrees latitude of los point lon (float): degrees longitude of los point enucoeffs (ndarray) size 3 array of the E,N,U coefficients of a line of sight vector. Comes from find_enu_coeffs. If this arg is used, others are not needed

Returns: ndarray: magnitudes same length as enu input, (k, 1)

Examples:

print('%.2f' % projectenuto_los([1,2,3],[1, 0, 0], 0, 0))

-2.00

print('%.2f' % projectenuto_los([1,2,3],[0, 1, 0], 0, 0))

-3.00

print('%.2f' % projectenuto_los([1,2,3],[0, 0, 1], 0, 0))

1.00

source

InsarLOS.rot — Method.

Find a 3x3 euler rotation matrix given an angle and axis.

Rotation matrix used for rotating a vector about a single axis.

Args: angle in degrees to rotate axis: 1, 2 or 3 for R1, R2, or R3 in_degrees (bool): specify the angle in degrees. if false, using radians

source

InsarLOS.rotate_xyz_to_enu — Method.

Rotates a vector in XYZ coords to ENU

Args: xyz: 3xk array of xyz vectors lat: latitude (deg) of point to rotate into lon: longitude (deg) of point to rotate into

Reference: https://gssc.esa.int/navipedia/index.php/TransformationsbetweenECEFandENU_coordinates

source